requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
django-filter
drf-spectacular
python-decouple
psycopg2-binary
gunicorn
whitenoise